Moving discussion-tool-selector directory
Adding it as a child of the pages-and-resources module.
