+++ src/index.scss
@@ -9,6 +9,6 @@
 
 @import "proctored-exam-settings/proctoredExamSettings.scss";
 
-@import "discussion-tool-selector/DiscussionToolSelector.scss";
+@import "pages-and-resources/discussions/DiscussionToolSelector.scss";
 
 @import "studio-header/header.scss";
