+++ src/pages-and-resources/discussions/discussion-tool-selector/features-table/FeaturesTable.jsx
